Dotnet FIT performer initial project setup
Motivation
----------
Adds the .NET solution and projects for the .NET FitProvider
implementation along with adding the C# style namespacing/package to the
.proto files.

Note that the solution requires that the
couchbaselabs/couchbase-transactions-dotnet be cloned in a directory
directly adjacent to this repo locally as it depends on the
Couchbase.Transactions project. In the future this can be changed, but
works for development.

Change-Id: Ief057c0ab4838afd6b991189d0dd4ce66ba0edf8
Reviewed-on: http://review.couchbase.org/c/transactions-fit-performer/+/131311
Reviewed-by: <richard.ponton@couchbase.com>
Reviewed-by: <praneeth.bokka@couchbase.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>