Refactoring so all Java GRPC classes are in same package
By default, the Java version of GRPC puts each message and enum
into a wrapper class based on the name of the .proto file.
So we have <package-name>.TxnClient.AttemptState for example.

I'm changing this so instead we have one Java file for each enum
and message, and they are called <package-name>.<enum-or-message-name>.

There's two reasons:
1. We end up with much smaller files that should cause less problems
   for the IDE.
2. We can freely move stuff between .proto files without it being
   a big refactoring chore.

There's also a small change to the proto files to allow AttemptStates
to be stored in TransactionCleanupAttempt, which will be needed by a
future test.  (The huge amount of code changes caused by this trivial
refactoring is what kicked off this larger changeset.)

Change-Id: I0c2c7c4f16ff9cf2edc672110557f6e98fdd75ab
Reviewed-on: http://review.couchbase.org/c/transactions-fit-performer/+/131369
Reviewed-by: <praneeth.bokka@couchbase.com>
Tested-by: Graham Pople <grahampople@gmail.com>