Driver, GRPC, Java performer: Improved validation of cleanup results
TransactionCleanupAttempt now passes some additional data that let us
quickly find the relevant ATR entry (speeding up tests), and paving
the way for some more advanced cleanup tests that are coming shortly.

This impacts a lot of lines as the cleanup validation routines needed
to change signature to support it, but is a low-impact change really.

Change-Id: Ibee780db0caea53331834a016793313f0435b8b8
Reviewed-on: http://review.couchbase.org/c/transactions-fit-performer/+/132288
Tested-by: Graham Pople <grahampople@gmail.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>