Preparing for custom metadata collection support by making GRPC pass around DocId references for ATRs
The GRPC protocol currently assumes that the ATR exists on the
default collection of a bucket - which it does, currently.

To support custom metadata collections, instead the protocol
will now consistently use a DocId (which contains full details
for a collection as well as a document id) to refer to an ATR.

The original ATR fields are kept for backwards compatibility,
but now marked as deprecated.  Performers should migrate to the
new fields.

A new DocAndColl class is added to FIT that maps to GRPC's DocId.

Change-Id: Ie7e406dacfb936ad337905f39fb260f7754ff19e
Reviewed-on: http://review.couchbase.org/c/transactions-fit-performer/+/138308
Reviewed-by: <praneeth.bokka@couchbase.com>
Tested-by: Graham Pople <grahampople@gmail.com>