Adding support for query to GRPC protocol and Java performer
This includes the new EXT_QUERY.

Note that query support requires relaxing the requirement that all
transaction operations can only raise a TransactionOperationFailed
error.  To support this, GRPC can now specify an exception
as the expected result of an operation.

In addition, the document status is no longer available in all cases,
e.g. when query is doing the get.  So no longer check this in all tests.
(Document status is only check by the Java performer anyway, and is
now deprecated.)

Change-Id: Ic1671645760554c199219d98990fbbe931f36f41
Reviewed-on: http://review.couchbase.org/c/transactions-fit-performer/+/140354
Tested-by: Graham Pople <grahampople@gmail.com>
Reviewed-by: <praneeth.bokka@couchbase.com>