CBD-3758: Remove FIT dependency on TransactionAttempt
In an effort to simplify the API surface area before C++ & .NET
GA releases, will be removing TransactionAttempt from the API.
This is returned in the TransactionsResult.attempts() field.

The only user of this information is likely to be FIT.  FIT
used it for two purposes:

* Verifying that what was written to the cluster matched what
  was returned in TransactionResult.  No longer needed if it's
  removed from TransactionResult.
* As a convenient shortcut to knowing where to look for the
  ATR entry.

For the second, ResultValidator is already polling all ATRs to
look for all attempts for the transaction, so this change should
not significantly impact test performance.  A similar polling
lookup needed to be added to a handful of tests that were
relying on the ATR information - especially cleanup-related tests.

A number of tests needed their validations adjusted, where the
in-memory account of attempts didn't match what was on-disc.

The Java performer no longer returns TransactionAttempt.

Change-Id: Idde0f4f47af54558b133f462c8bee3ac792f31d7
Reviewed-on: http://review.couchbase.org/c/transactions-fit-performer/+/140371
Tested-by: Graham Pople <grahampople@gmail.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>