GRPC and Java performer adding features for ExtMobileInterop
Detecting the absence of events, which is validation
performed on the performer side.

Adding EC_ANYTHING, as part of the ErrorClass
deprecation.  But performers instead of handling this
should just remove ErrorClass validation.

Enhancing stashing so that instead of a single stashed
result - which is not compatible with multiple or concurrent
operations - the performer maintains a per-transaction map
of 'stash slots' to TransactionGetResults.

Adding EXT_MOBILE_INTEROP and accompanying hooks.

All features need to be implemented by performers implementing
EXT_MOBILE_INTEROP.  The driver will not use these features
with performers that do not declare support for that.

Java performer changes checked in simultaneously as a
reference implementation.

Change-Id: Ic559d11267f2662b98f69bb549dd0b613ded45b3
Reviewed-on: https://review.couchbase.org/c/transactions-fit-performer/+/173491
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: <praneeth.bokka@couchbase.com>