Continuing ExtSingleQuery rework
If the transaction expires, a single query transaction
should raise an UnambiguousTimeoutException rather than
a TransactionExpired.

Also tweaking how a non-TransactionFailed error is
indicated in the GRPC, to make it clearer.

Creating required indexes in single query tests.

Renaming a test so it fits the naming of the other files.

And making all single query tests TestInstance.Lifecycle.PER_CLASS,
so their constructor is only run once.

Change-Id: Ibc48cd732c5711823df3c0bbfddccc1578f64242
Reviewed-on: https://review.couchbase.org/c/transactions-fit-performer/+/173678
Reviewed-by: <matthew.bray@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>