Major refactoring of the GRPC
As we now have SDK and transactions being tested by FIT, I'm
continuing to try and find the best way of structuring the
GRPC accordingly.

A lot of messages have been renamed.  The top-level rpc is now
`run`, which sends a run.Request and gets back a stream of
run.Responses.  A run.Request contains a `run.Workloads`.  Many
other messages have been renamed or tidied (but nothing used by
existing performers).

See the README for the rules now followed here.
This is a breaking change, by necessity of rule 5.
All performers will be patched in a follow-up commit.

Other important changes:
- Adds WIP initial support for KV range scans.
- In particular this requires WIP initial support for generic
  streaming, which we can also use for query, analytics etc.

Change-Id: Ic05caab1f3a783867e95b48f4dbe3edca1856755
Reviewed-on: https://review.couchbase.org/c/transactions-fit-performer/+/177900
Tested-by: Graham Pople <graham.pople@couchbase.com>
Reviewed-by: <matthew.bray@couchbase.com>
Reviewed-by: <praneeth.bokka@couchbase.com>