SDKQE-2745: Support FIT performers that are not ExtSDKIntegrated but do have performerCapsFetch
At the same time as ExtSDKIntegration, new rpcs were added
(performerCapsFetch, clusterConnectionCreate and clusterConnectionClose)
as a better version of createConnection.

However these new rpcs were not intended to be bound to ExtSDKIntegration.
A non-integrated performer should be able to implement them too.

This changeset now allows that, and adds a rule in the GRPC documentation
about implementing all 3 if any are implemented, to simplify the test space.

Change-Id: Ifab6610d29740dc31d4ac2fb7da86d08065baf9e
Reviewed-on: https://review.couchbase.org/c/transactions-fit-performer/+/178635
Reviewed-by: <matthew.bray@couchbase.com>
Tested-by: Graham Pople <graham.pople@couchbase.com>