FIT GRPC - moving and refactoring QueryIndexManager
With Serverless it's becoming increasingly common to need
similar interfaces at the Cluster, Bucket Scope and/or
Collection level.

This changeset proposes a naming and message scheme that
represents this cleanly and that we can apply consistently
going forwards.

It also migrates the QueryIndexManager and
CollectionQueryIndexManager GRPC to this new scheme.  This
is a breaking change and the affected performers will need
changing.

It also finds a cleaner, more DRYed, representation of that
GRPC.

Change-Id: Ifc0d453595b2f347cd455e93958249fbea308f24
Reviewed-on: https://review.couchbase.org/c/transactions-fit-performer/+/187131
Tested-by: Graham Pople <graham.pople@couchbase.com>
Reviewed-by: Will Broadbelt <william.broadbelt@couchbase.com>