SDKQE-3439: Columnar - Support all JSON types in ContentWas
Motivation
==========
In SDKs with dynamic row typing, ContentAs is ignored and the driver has no control over what type the rows will be returned in. We should support all possible types for JSON values in ContentWas so performers can always return the rows in the type they were decoded in.

Changes
=======
* Add boolean, string, double, integer, null types in ContentWas
* Add a simple test that returns a string row

Change-Id: I5bec7091f89c62c3cf3f8bd6cb48da9de3325bf7
Reviewed-on: https://review.couchbase.org/c/transactions-fit-performer/+/217245
Tested-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>
Reviewed-by: Will Broadbelt <william.broadbelt@couchbase.com>