Update Columnar driver to support Enterprise Analytics
Add a flag to the performer capability response
indicating whether the performer is for Columnar or Analytics.

In testClusterConnectionWithInvalidHostname(), create the
modified connection string using the same scheme as the
performer connection string, instead of assuming `couchbases`.

In FitConfig.performerConnectionIsCouchbase2, first check
for an HTTP scheme, otherwise ClusterConnection.scheme throws
an exception (HTTP is not one of the schemes supported by
SDK 3's ConnectionString.Scheme.

When the proxy feeds a connection string to the performer,
use a scheme appropriate to the performer type;
"couchbases" for Columnar, "https" for Analytics.

Change-Id: Ib49e305e6283130779526b4afbe643d51db2dc57
Reviewed-on: https://review.couchbase.org/c/transactions-fit-performer/+/226812
Tested-by: David Nault <david.nault@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>